+++ bin/src/pubspec.dart
@@ -7,7 +7,8 @@ void insertPubspec({required ArgResults from}) {
   bool insert = from["pubspec"] ?? true;
   String input = from["input"]!.toString();
   if (insert) {
-    String output = from["output"]!.toString().replaceAll(RegExp(r'/[^/]+$'), "/");
+    String output =
+        from["output"]!.toString().replaceAll(RegExp(r'/[^/]+$'), "/");
     if (output.startsWith("assets/")) {
       File pubspec = File('pubspec.yaml');
       List<String> lines = pubspec.readAsLinesSync();
@@ -21,8 +22,10 @@ void insertPubspec({required ArgResults from}) {
       }
 
       if (!existed) {
-        int flutterIndex = lines.lastIndexWhere((line) => line.trim() == 'flutter:');
-        int assetsIndex = lines.lastIndexWhere((line) => line.trim() == 'assets:');
+        int flutterIndex =
+            lines.lastIndexWhere((line) => line.trim() == 'flutter:');
+        int assetsIndex =
+            lines.lastIndexWhere((line) => line.trim() == 'assets:');
         if (flutterIndex == -1) {
           lines.insert(lines.length - 1, 'flutter:\n  assets:\n    - $output');
         } else {
@@ -36,8 +39,10 @@ void insertPubspec({required ArgResults from}) {
         pubspec.writeAsStringSync(lines.join('\n'));
       }
     } else {
-      print("\u001b[33m--pubspec\u001b[0m \u001b[2mflag ignored, due to output not in assets directory\u001b[0m");
-      print("\u001b[2mThis make the $input not accesible for every flutter platform\u001b[0m");
+      print(
+          "\u001b[33m--pubspec\u001b[0m \u001b[2mflag ignored, due to output not in assets directory\u001b[0m");
+      print(
+          "\u001b[2mThis make the $input not accesible for every flutter platform\u001b[0m");
     }
   }
 }